backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
psycopg2-binary
alembic
pydantic-settings
redis
celery[redis]
pandas
numpy
scikit-learn
xgboost
lightgbm
# catboost
# torch
optuna
joblib
shap
python-dotenv
httpx
gunicorn
loguru
requests
beautifulsoup4
pytest
passlib[bcrypt]
python-jose[cryptography]
python-multipart
